back/requirements.txt
absl-py==2.3.1
asgiref==3.10.0
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
idna==3.11
keras==3.13.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mido==1.3.3
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
# opencv-python==4.12.0.88
opencv-python-headless
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-rtmidi==1.5.8
pytz==2025.2
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.20.0
termcolor==3.3.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
Werkzeug==3.1.4
wheel==0.45.1
whitenoise==6.11.0
wrapt==2.0.1
dj-database-url==1.2.0